moved Req/Res interfaces to shared-dto
