--- packages/shared-dto/index.ts
@@ -1,3 +1,0 @@
-export interface QueryPayload {
-  payload: string;
-}
