--- packages/shared-dto/src/User.ts
@@ -1,7 +1,0 @@
-interface User {
-  place: number;
-  name: string;
-  email: string;
-  phoneNumber: string; // maybe make separate type for this
-  timeJoined: number;
-}
